Interactively Create Rmarkdown Chunks

Addin that splits full documents into Rmarkdown chunks or wraps highlighted text in an active RStudio document editor in a Rmarkdown chunk.

chunkr()

chunkr_doc()

chunkr_section()

Details

Setting remedy_opts$set(list(full_doc=TRUE)) chunkr will convert entire document by remedy_opts$get('token_purl'). The default token is the output from a purl conversion (## ----chunk name,chunk options----).

Setting remedy_opts$set(list(full_doc=FALSE)), user highlights text and chunkr will wrap a new Rmarkdown chunk around it, utilizing the other options in remedy_opts.

Examples

# NOT RUN {
remedy_example(
    c( "# Create a chunk",
    "a <- 12",
    "aaa <- 13"),
    chunkr
    )
# }