Default and current remedy options

Options for functions in the remedy package. When running R code, the object remedy_opts (default options) is not modified by chunk headers (local chunk options are merged with default options), whereas remedy_opts_current (current options) changes with different chunk headers and it always reflects the options for the current chunk.

Normally we set up the global options once in the first code chunk in a document using remedy_opts$set(), so that all latter chunks will use these options. Note the global options set in one chunk will not affect the options in this chunk itself, and that is why we often need to set global options in a separate chunk.

Below is a list of default chunk options, retrieved via remedy_opts$get():

remedy_opts

remedy_opts_current

Format

An object of class list of length 5.

An object of class list of length 5.

Note

remedy_opts_current is read-only in the sense that it does nothing if you call remedy_opts_current$set(); you can only query the options via remedy_opts_current$get().

Examples

remedy_opts$get()

#> $basic
#> [1] FALSE
#> 
#> $name
#> [1] "remedy"
#> 
#> $counter
#> [1] TRUE
#> 
#> $chunk_opts
#> NULL
#> 
#> $kable_opts
#> NULL
#> 
#> $full_doc
#> [1] FALSE
#> 
#> $token_purl
#> [1] "^#{2} -{4}(.*?)-{4,}$"
#> 
#> $token_url
#> [1] "^(?:(?:https?|ftp|file)://|www\\.|ftp\\.)[A-z0-9+&@#/%=~_|$?!:,.-]*[A-z0-9+&@#/%=~_|$]$"
#> 
#> $token_rel_link
#> [1] "^.*[/|\\.][^\\.]+$"
#> 
#> $token_img_link
#> [1] "jpeg" "jpg"  "png"  "gif" 
#> 
#> $youtube_output
#> [1] "html"
#> 
#> $youtube_width
#> [1] "100%"
#> 
#> $youtube_height
#> [1] "400"
#> 
#> $hotkeys
#>           backtick               bold              chunk         chunksplit 
#>       "Ctrl+Cmd+`"       "Ctrl+Cmd+B"   "Ctrl+Alt+Cmd+C" "Ctrl+Shift+Alt+C" 
#>          chunkname           footnote                 h1                 h2 
#> "Ctrl+Shift+Alt+N" "Ctrl+Cmd+Shift+6"       "Ctrl+Cmd+1"       "Ctrl+Cmd+2" 
#>                 h3                 h4                 h5                 h6 
#>       "Ctrl+Cmd+3"       "Ctrl+Cmd+4"       "Ctrl+Cmd+5"       "Ctrl+Cmd+6" 
#>        htmlcomment              image            italics              latex 
#>       "Ctrl+Alt+C"       "Ctrl+Cmd+P"       "Ctrl+Cmd+I"       "Ctrl+Cmd+L" 
#>               list              right             strike              table 
#> "Ctrl+Shift+Cmd+="    "Alt+Cmd+Right"       "Ctrl+Cmd+S"       "Ctrl+Cmd+T" 
#>                url           xaringan            youtube 
#>       "Ctrl+Cmd+U"       "Ctrl+Cmd+X"       "Ctrl+Cmd+Y" 
#>